refactor: export pages for folder pages
