--- client/src/pages/main.tsx
@@ -1,7 +1,0 @@
-export const MainPage = () => {
-  return (
-    <div>
-      Main page
-    </div>
-  )
-}
