+++ client/src/pages/main/ui/main.tsx
@@ -1,0 +1,3 @@
+export const MainPage = () => {
+  return <div>Main page</div>;
+};
