+++ client/src/pages/map/index.ts
@@ -1,0 +1,1 @@
+export { MapPage } from './ui/map';
