+++ client/src/pages/profile-customer/index.ts
@@ -1,0 +1,1 @@
+export { ProfileCustomerPage } from './ui/profile-customer';
