+++ client/src/pages/profile-vendor/index.ts
@@ -1,0 +1,1 @@
+export { ProfileVendorPage } from './ui/profile-vendor';
