+++ client/src/pages/register/index.ts
@@ -1,0 +1,1 @@
+export { RegisterPage } from './ui/register';
